sentiment_analysis/test_requirements.txt
numpy==1.19.5
transformers==4.1.1
scikit-learn==0.24.0
tqdm==4.56.0
pandas==1.1.5
pyyaml==5.4.1
xlrd==1.1.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cpu
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.8.2+cpu
